misc: restructure folder hierarchy
