+++ .gitmodules
@@ -1,6 +1,3 @@
-[submodule "external/catch2"]
-	path = external/catch2
+[submodule "catch2"]
+	path = cpp/tests/catch2
 	url = https://github.com/catchorg/Catch2.git
-[submodule "external/md5"]
-	path = external/md5
-	url = https://github.com/Wodann/constexpr-md5-cpp.git
