--- include/mun/mun.h
@@ -1,9 +1,0 @@
-#ifndef MUN_MUN_H_
-#define MUN_MUN_H_
-
-#include "mun/error.h"
-#include "mun/invoke_fn.h"
-#include "mun/runtime.h"
-#include "mun/struct_ref.h"
-
-#endif
