--- include/mun/util.h
@@ -1,10 +1,0 @@
-#ifndef MUN_UTIL_H_
-#define MUN_UTIL_H_
-
-#if defined(__clang__) || defined(_MSC_VER)
-#define MUN_CALLTYPE __cdecl
-#else
-#define MUN_CALLTYPE
-#endif
-
-#endif
